Probation casein leave (#66)
* leave status issue resolution

* sending context to master

* context and leave status changes

* model duplication removal

* EHRM context updated

* entity push in

* probation case handled in leave

---------

+++ EHRM.DAL/Database/LeaveBalance.cs
@@ -8,7 +8,7 @@ public partial class LeaveBalance
 
     public int EmpId { get; set; }
 
-    public int TenureYears { get; set; }
+    public decimal TenureYears { get; set; }
 
     public int EarnedLeave { get; set; }
