+++ EHRM.ViewModel/Leave/LeaveBalanceViewModel.cs
@@ -10,7 +10,7 @@ namespace EHRM.ViewModel.Leave;
 public class LeaveBalanceViewModel
 {
     public int EmpId { get; set; }
-    public int TenureYears { get; set; }
+    public decimal TenureYears { get; set; }
     public int EarnedLeave { get; set; }
     public int SickLeave { get; set; }
     public int CasualLeave { get; set; }
